New greynoise integration (#1204)
* optimized Dockerfile

* pushed back certego-ui last version

* 1192 greynoise updates (#1193)

* Bump django from 3.2.14 to 3.2.15 in /requirements (#1144)

Bumps [django](https://github.com/django/django) from 3.2.14 to 3.2.15.
- [Release notes](https://github.com/django/django/releases)
- [Commits](https://github.com/django/django/compare/3.2.14...3.2.15)

---
updated-dependencies:
- dependency-name: django
  dependency-type: direct:production
...



* greynoise updates

fix requirement mismatch

rename analyzer file to remove conflict with SDK

handle exceptions from SDK

move exception to f string

greynoise updates

fix required flag


* revert certego ui

* certego-ui 0.1.3

* fix dockerfile

* add chance to allow to do not use api key for community endpoint

* adjust

* test

* test

* test

* test

* test

* test

* test refactor monkey patch

* Revert "test refactor monkey patch"

This reverts commit 32aba47493ccdd228258f54d7bf82c738ef77a2a.
