+++ requirements/project-requirements.txt
@@ -64,6 +64,7 @@ speakeasy-emulator==1.5.9
 telfhash==0.9.8
 yara-python==4.2.0
 pydragonfly==0.1.0
+greynoise==1.3.0
 XLMMacroDeobfuscator[secure]==0.2.3
 # this is required because XLMMacroDeobfuscator does not pin xlrd2 and pyxlsb2 packages
 pyxlsb2==0.0.8
